sessions/session-01-ai-chatbots/participants/Beloved1310/code/requirements.txt
google-cloud-aiplatform>=1.26.0
vertexai>=0.1.0
python-dotenv>=1.0.0
streamlit>=1.28.0
requests>=2.31.0
google-generativeai>=0.3.0
beautifulsoup4==4.14.2
requests==2.32.5
